KAFKA-7616; Make MockConsumer only add entries to the partition map returned by poll() if there are any records to return
…eturned by poll() if there are any records to return

The MockConsumer behaves unlike the real consumer in that it can return a non-empty ConsumerRecords from poll, that also has a count of 0. This change makes the MockConsumer only add partitions to the ConsumerRecords if there are records to return for those partitions.

A unit test in MockConsumerTest demonstrates the issue.

Author: Stig Rohde Døssing <stigdoessing@gmail.com>

Reviewers: Manikumar Reddy <manikumar.reddy@gmail.com>

Closes #5901 from srdo/KAFKA-7616