MINOR: Fix flaky ConsumerTopicCreationTest (#6727)
`ConsumerTopicCreationTest` relied on `KafkaConsumer#poll` to send a `MetadataRequest` within 100ms to verify if a topic is auto created or not. This is brittle and does not guarantee if the request made it to the broker or was processed successfully. This PR fixes the flaky test by adding another topic; we wait until we consume a previously produced record to this topic. This ensures MetadataRequest was processed and we could then check if the topic we're interested in was created or not.

Reviewers: Boyang Chen <bchen11@outlook.com>, Jason Gustafson <jason@confluent.io>