MINOR: Remove spammy log message during topic deletion
Deletion of a large number of topics can cause a ton of log spam. In a test case on 2.2, deletion of 50 topics with 100 partitions each caused about 158 Mb of data in the controller log. With the improvements to batch StopReplica and the patch here, we reduce that to about 1.5 Mb.

Kudos to gwenshap for spotting these spammy messages.

Author: Jason Gustafson <jason@confluent.io>

Reviewers: Gwen Shapira

Closes #6738 from hachikuji/remove-verbose-topic-deletion-log-message